+++ src/capturer/model.rs
@@ -1,11 +1,15 @@
 use crate::model::Result;
 use windows::Win32::Graphics::Dxgi::{
-  DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO, DXGI_OUTPUT_DESC,
+  DXGI_OUTDUPL_DESC, DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO, DXGI_OUTPUT_DESC,
 };
 
 /// Capturer is stateful, it holds a buffer of the last captured frame.
 pub trait Capturer {
+  /// This is usually used to get the screen's position and size.
   fn dxgi_output_desc(&self) -> Result<DXGI_OUTPUT_DESC>;
+
+  /// This is usually used to get the screen's pixel width/height and buffer size.
+  fn dxgi_outdupl_desc(&self) -> DXGI_OUTDUPL_DESC;
 
   /// Get the buffer of the last captured frame.
   /// The buffer is in BGRA32 format.